refactor: move consent utilities to consent-constants.ts
- Move hasLocalConsent() and setLocalConsent() to consent-constants.ts
- Export ConsentData interface from consent-constants
- Import utilities from consent-constants in CookieBanner
